DRY up participant action services
Add `Participants::Action` abstract class that individual actions can inherit
from. Move specs into a shared context.
