+++ app/services/participants/defer.rb
@@ -1,10 +1,7 @@
 # frozen_string_literal: true
 
 module Participants
-  class Defer
-    include ActiveModel::Model
-    include ActiveModel::Attributes
-
+  class Defer < Action
     DEFERRAL_REASONS = %w[
       bereavement
       long-term-sickness
@@ -13,16 +10,9 @@ module Participants
       other
     ].freeze
 
-    attribute :lead_provider
-    attribute :participant
-    attribute :course_identifier
     attribute :reason
 
-    validates :lead_provider, presence: true
-    validates :participant, presence: true
-    validates :course_identifier, inclusion: { in: Course::IDENTIFIERS }, allow_blank: false
     validates :reason, inclusion: { in: DEFERRAL_REASONS }, allow_blank: false
-    validate :application_exists
     validate :not_already_deferred
     validate :not_withdrawn
     validate :has_declarations
@@ -31,7 +21,7 @@ module Participants
       return false if invalid?
 
       ActiveRecord::Base.transaction do
-        create_application_state!
+        create_application_state!(state: :deferred, reason:)
         application.deferred!
         participant.reload
       end
@@ -40,22 +30,6 @@ module Participants
     end
 
   private
-
-    def create_application_state!
-      ApplicationState.create!(application:, lead_provider:, reason:, state: :deferred)
-    end
-
-    def application
-      @application ||= participant
-        &.applications
-        &.accepted
-        &.includes(:course)
-        &.find_by(lead_provider:, course: { identifier: course_identifier })
-    end
-
-    def application_exists
-      errors.add(:participant, :blank) if application.blank?
-    end
 
     def not_withdrawn
       errors.add(:participant, :already_withdrawn) if application&.withdrawn?
