+++ app/services/participants/resume.rb
@@ -1,18 +1,7 @@
 # frozen_string_literal: true
 
 module Participants
-  class Resume
-    include ActiveModel::Model
-    include ActiveModel::Attributes
-
-    attribute :lead_provider
-    attribute :participant
-    attribute :course_identifier
-
-    validates :lead_provider, presence: true
-    validates :participant, presence: true
-    validates :course_identifier, inclusion: { in: Course::IDENTIFIERS }, allow_blank: false
-    validate :application_exists
+  class Resume < Action
     validate :not_already_active
 
     def resume
@@ -28,22 +17,6 @@ module Participants
     end
 
   private
-
-    def create_application_state!
-      ApplicationState.create!(application:, lead_provider:)
-    end
-
-    def application
-      @application ||= participant
-        &.applications
-        &.accepted
-        &.includes(:course)
-        &.find_by(lead_provider:, course: { identifier: course_identifier })
-    end
-
-    def application_exists
-      errors.add(:participant, :blank) if application.blank?
-    end
 
     def not_already_active
       errors.add(:participant, :already_active) if application&.active?
